+++ packages/api/resolvers/mutations/setPassword.js
@@ -1,6 +1,6 @@
 import { log } from 'meteor/unchained:core-logger';
 import { Users } from 'meteor/unchained:core-users';
-import { UserNotFoundError } from '../../errors';
+import { UserNotFoundError, InvalidIdError } from '../../errors';
 
 export default async function setPassword(
   root,
@@ -9,6 +9,7 @@ export default async function setPassword(
 ) {
   log(`mutation setPassword ${foreignUserId}`, { userId: ownUserId });
   const userId = foreignUserId || ownUserId;
+  if (!userId) throw new InvalidIdError({ userId });
   const user = Users.findOne({ _id: userId });
   if (!user) throw new UserNotFoundError({ userId });
   const res = await user.setPassword(newPassword);
